backend/requirements.txt
# FastAPI and ASGI
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
asyncpg==0.29.0
psycopg2-binary>=2.9.9

# Redis
redis==5.0.1

# Celery
celery==5.3.6

# Pydantic
pydantic==2.5.3
pydantic-settings==2.1.0

# Milvus
pymilvus==2.3.6

# MinIO
minio==7.2.3

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# HTTP Client
httpx==0.26.0
aiohttp==3.9.1

# MongoDB (for FastGPT sync)
motor==3.3.2
pymongo==4.6.1

# Evaluation Frameworks
ragas==0.1.7
datasets==2.16.1
openai==1.8.0
langchain-openai>=0.1.0

# Document Processing
pypdf>=4.0.0
python-docx>=1.1.0

# Data Processing
pandas>=2.1.4
numpy==1.26.3

# Utilities
python-dotenv==1.0.0
orjson==3.9.12
httpx==0.26.0

# Development
pytest==7.4.4
pytest-asyncio==0.23.3